refactor code to earth2grid.healpix.coordinates
